% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFiles.R
\name{combineSectionsByMarkerLayer}
\alias{combineSectionsByMarkerLayer}
\title{Heuristically determine varve sequence order by marker layers}
\usage{
combineSectionsByMarkerLayer(sortedSequence)
}
\arguments{
\item{sortedSequence}{a list of varve data.frames in stratigraphic order}
}
\value{
A data.frame of the full varve sequence.
}
\description{
combines sections into sequence
}
\author{
Nick McKay
}
