% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varveModel.R
\name{createCoreInputForModel}
\alias{createCoreInputForModel}
\title{create cores data.frame for varve model}
\usage{
createCoreInputForModel(ensOut, ocPrior = 0.05, ucPrior = 0.05,
  mlPrior = 0.8)
}
\arguments{
\item{ensOut}{output of generateThicknessEnsemble()}

\item{mlPrior}{}
}
\description{
create cores data.frame for varve model
}
