% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateOverAndUndercounting.R
\name{createTableChunks}
\alias{createTableChunks}
\title{create table chunks to insert into a composite sequence}
\usage{
createTableChunks(rowInd, simVarves, compSeq)
}
\arguments{
\item{rowInd}{row index for chunks to create}

\item{simVarves}{the simulated varves for each chunk}

\item{compSeq}{the composite sequence to work with}
}
\description{
create table chunks to insert into a composite sequence
}
