% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFiles.R
\name{determineMarkerLayerOrder}
\alias{determineMarkerLayerOrder}
\title{Heuristically determine varve sequence order by marker layers}
\usage{
determineMarkerLayerOrder(sequence)
}
\arguments{
\item{sequence}{a list of varve data.frames in any arbitrary order}
}
\value{
A vector of indices with the sequence order
}
\description{
Tries to determine the order of the sequence based on the marker layers
}
\author{
Nick McKay
}
