% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateOverAndUndercounting.R
\name{emulateVarveThickness}
\alias{emulateVarveThickness}
\title{simulate varve thickness for an uncounted(able) section}
\usage{
emulateVarveThickness(totalThickness, exampleVarves)
}
\arguments{
\item{totalThickness}{the total thickness of the section}

\item{exampleVarves}{a (ideally long) series of varves to inform the emulation}
}
\value{
emulated varves
}
\description{
simulate varve thickness for an uncounted(able) section
}
