% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateOverAndUndercounting.R
\name{generateThicknessEnsemble}
\alias{generateThicknessEnsemble}
\title{Generate thickness ensemble}
\usage{
generateThicknessEnsemble(compSeq, nEns = 1000)
}
\arguments{
\item{compSeq}{a data.frame of composited sequence}

\item{nEns}{number of ensembles}
}
\value{
a matrix of thickness ensemble members
}
\description{
Generate thickness ensemble
}
