% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotSections}
\alias{plotSections}
\title{Plot a bunch of varve sections}
\usage{
plotSections(varveSequenceList, output = NA)
}
\arguments{
\item{output}{A pdf file to concatenate the output into}
}
\value{
a
}
\description{
Plot a bunch of varve sections
}
