% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFiles.R
\name{readVarveDirectory}
\alias{readVarveDirectory}
\title{Read multiple varve sequences thicknesses from a directory of shape files}
\usage{
readVarveDirectory(directory = NULL, codeCol = c("conf", "VarveID"),
  varveTop = "top", markerCol = c("Marker", "Markers"),
  scaleToThickness = NA)
}
\arguments{
\item{directory}{the path to to the directory. Leaving it blank will open a file chooser.}

\item{codeCol}{name of the column of varve confidence ratings. Can be a single and or multiple strings in a vector. Default = c("conf","VarveID")}

\item{varveTop}{where is the top of the varve sequence on the shape file. Options are "top","bottom","left","right". Default = "top".}

\item{markerCol}{name of the column of marker layer names. Can be a single and or multiple strings in a vector. Default = c("Marker","Markers")}

\item{scaleToThickness}{Value to scale the sum of the measurements. NA retains the scale in the .shp file.}
}
\value{
A list of data.frames with the varve data
}
\description{
Reads in a directory of varve thicknesses files, using readVarveShapefile.
}
\author{
Nick McKay
}
