% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateOverAndUndercounting.R
\name{simulateOverAndUndercounting}
\alias{simulateOverAndUndercounting}
\title{simulate over and under counting}
\usage{
simulateOverAndUndercounting(compSeq)
}
\arguments{
\item{compSeq}{a composite sequence data.frame}
}
\value{
simulate counts
}
\description{
simulate over and under counting
}
