% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateOverAndUndercounting.R
\name{translateCodesToProbabilities}
\alias{translateCodesToProbabilities}
\title{Translate varve codes to over and undercounting probabilities}
\usage{
translateCodesToProbabilities(compSeq, translationTable, baselineProb = 0.05)
}
\arguments{
\item{compSeq}{a composite sequence data.frame}

\item{translationTable}{a data.frame that translates varve codes to over and undercountin probabilities}

\item{baselineProb}{what baseline probability should be used in the absence of varve cides}
}
\description{
Translate varve codes to over and undercounting probabilities
}
