% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFiles.R
\name{readVarveShapefile}
\alias{readVarveShapefile}
\title{Read varve thicknesses from shape file}
\usage{
readVarveShapefile(filename, codeCol = "conf")
}
\arguments{
\item{filename}{the path to a shape file}
}
\value{
A data.frame with the varve data
}
\description{
Pulls varve thicknesses and error codes from a shapefile counted in GIS
}
\author{
Nick McKay
}
